pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "select-bq"
version = "0.1.1"
description = "Select BigQuery CLI - SELECT-only queries with allowlist and audit logging"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["bigquery", "cli", "security", "gcp", "sql", "allowlist", "audit"]
authors = [{ name = "select-bq contributors" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
]
dependencies = [
    "sqlglot>=24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0",
    "twine>=5.0",
]

[project.urls]
Homepage = "https://github.com/sqlhabit/select-bq"
Repository = "https://github.com/sqlhabit/select-bq"

[project.scripts]
select-bq = "select_bq.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
